+++ src/test/scala/com/dwolla/lambda/cloudflare/record/CloudflareDnsRecordHandlerSpec.scala
@@ -84,17 +84,6 @@ object CloudflareQueryParams {
 
 }
 
-class DnsRecordClientStub[F[+_]](const: F[Nothing]) extends DnsRecordClient[F] {
-  override def getById(zoneId: ZoneId, resourceId: ResourceId): F[IdentifiedDnsRecord] = const
-  override def createDnsRecord(record: UnidentifiedDnsRecord): F[IdentifiedDnsRecord] = const
-  override def updateDnsRecord(record: IdentifiedDnsRecord): F[IdentifiedDnsRecord] = const
-  override def getExistingDnsRecords(name: String, content: Option[String], recordType: Option[String]): F[IdentifiedDnsRecord] = const
-  override def deleteDnsRecord(physicalResourceId: String): F[PhysicalResourceId] = const
-  @targetName3("deleteDnsRecordNewtype")
-  override def deleteDnsRecord(physicalResourceId: PhysicalResourceId): F[PhysicalResourceId] = const
-  override def getByUri(uri: String): F[IdentifiedDnsRecord] = const
-}
-
 @annotation.experimental
 class UpdateCloudflareSpec extends CatsEffectSuite {
